requirements.txt
requests>=2.31.0
openai>=1.0.0
gspread>=5.12.0
google-auth>=2.23.0
